move tests into separate folders and split where necessary
